Release: 1.0.3 (#20)
* Bumped up version to 1.0.3 and updated change log

This commit ups the version of project ChatGPT from 1.0.2 to 1.0.3 in pom.xml. The change log in README.md is also updated with the new version's details. This latest version includes fixes to the issue #17 previously reported on GitHub.

* Update dependencies in pom.xml

Updated the version number of hutool-core from 5.8.10 to 5.8.11 and apache httpcomponents from 4.5.5 to 4.5.13 in pom.xml. The upgrade was done to include the latest performance improvements and bug fixes present in the newer versions of these libraries.